Added function mu4e-dashboard to open the default dashboard file
Currently, mu4e-dashboard-mode operates in the current file, but the keymaps are always
loaded from the file mu4e-dashboard-file (it silently does nothing if the file does
not exist)

This means that, when the current file is not mu4e-dashboard-file, the file's keybindings
are ignored. This bug is fixed in this commit.

Also, a new function is added: mu4e-dashboard, that will visit mu4e-dashboard-file and
run mu4e-dashboard-mode on it. It will generate an error if the file does not exist.